+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `state_vector_simulation.vec_simulation` accepts a user specified `initial_state` now.
 
 ## [0.2.0] - 2022-01-11
 ### Changed